\name{LRVBUtils-package}
\alias{LRVBUtils-package}
\alias{LRVBUtils}
\docType{package}
\title{
\packageTitle{LRVBUtils}
}
\description{
\packageDescription{LRVBUtils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LRVBUtils}
\packageIndices{LRVBUtils}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LRVBUtils}

Maintainer: \packageMaintainer{LRVBUtils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
